s3-param Remove .offset == 0 checks as 'valid' will have offset 0
The validity of an entry in the parm_table is by having a .p_class of
P_LOCAL or P_GLOBAL rather than P_SEPARATOR.  Termination of the table
is by having a .label of non-NULL.  This is possible because there are
no longer any specially handled smb.conf options without a value in
the struct loadparm_globals or struct loadparm_service.

This is required because the first element in the structure will have
.offset = 0, and skipping that would be unfortunate (particularly as it
is the vital 'valid' variable).

Andrew Bartlett