IP allocation simulation - add -H/-S options for hard/soft imbalance limit.
An imbalance exceeding the hard limit, as specified by -H (and
defaulting to 1), now causes termination when -x is specified.

Imbalances exceeding the soft limit, as specified by -S (and
defaulting to 1), are counted and printed in the statistics summary.

A side-effect is that imbalances less than 2 are no longer rounded
down to 0, since we want to see them in the stats.

Signed-off-by: Martin Schwenke <martin@meltin.net>

(This used to be ctdb commit b5e9a4c50eedb8cc786c52af06352788ca25f51e)