IP allocation simulation - add analysis of IP groups.
The public addresses passed to the node constructer can be nested 2
levels.  Each sub-list is an IP group for which separate balance
analysis is done.  However, the public address list is flattened and
the actual IP assignment algorithm doesn't know about IP groups.

This allows extra statistics to be printed and an extra termination
condition to be added for unbalanced IP groups.

Most code from calculate_imbalance() is factored out to a a new
function imbalance_for_ips(), which calculates imbalance for the given
IPs.  calculate_imbalance() now returns the overall imbalance and a
list containing imbalances for each IP group.  To support this
node_ip_coverage() now takes an optional list of IPs to check coverage
within.

This also adds extra output to show statistics for the LCP2 imbalance
metric.

Signed-off-by: Martin Schwenke <martin@meltin.net>

(This used to be ctdb commit 701395087156b2a5c7be1564897b796df35b69ec)