IP allocation simulation - add -A/--aggressive option.
This is likely to cause many more state changes for nodes.  In this
mode the odds of a failover are applied to determine whether a state
change occurs for each node.  If no state change occurs then the
process is repeated.

Signed-off-by: Martin Schwenke <martin@meltin.net>

(This used to be ctdb commit b7c42bff9457ec8294b04245af8e3b6010707d1a)