nbt: fix WinXP S3 domain join: alignment of nbt_netlogon_response_from_pdc
This fixes WinXP joining a Samba3 domain, which was broken on hosts
with an even number of characters in the host name.  The alignment
requested in the structure was ignored because of the overall
NDR_NOALIGN set on the packet.

Andrew Bartlett

Autobuild-User: Jeremy Allison <jra@samba.org>
Autobuild-Date: Thu Jul 28 00:30:26 CEST 2011 on sn-devel-104